New shit

+++ workforce-backend/src/modules/scheduling/shiftTemplate/shiftTemplate.service.ts
@@ -1,3 +1,4 @@
+import Schedule from "../../../models/schedule.model";
 import ShiftTemplate from "../../../models/shiftTemplate.model";
 
 export const createTemplate = async (data: any) => {
@@ -17,5 +18,11 @@ export const updateTemplate = async (id: string, data: any) => {
 };
 
 export const deleteTemplate = async (id: string) => {
+  const assignedSchedule = await Schedule.findOne({ shiftTemplateId: id });
+
+  if (assignedSchedule) {
+    throw new Error("Cannot delete a template assigned to schedules");
+  }
+
   return await ShiftTemplate.findByIdAndDelete(id);
 };
